# Speed up rollup module build process